Moved some Vote classes to new package, created new VoteService and Repository.
