--- src/main/java/com/eacuamba/dev/reddit_clone_using_angular_spring/domain/model/Vote.java
@@ -1,33 +1,0 @@
-package com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
-
-import java.time.Instant;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-
-@Table
-@Entity
-public class Vote {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "vote_type")
-    @Enumerated(EnumType.STRING)
-    private VoteType voteType;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", referencedColumnName = "id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", referencedColumnName = "id")
-    private User user;
-}
