+++ src/main/java/com/eacuamba/dev/reddit_clone_using_angular_spring/domain/model/vote/Vote.java
@@ -1,0 +1,54 @@
+package com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.vote;
+
+import com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.Post;
+import com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.User;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+import org.hibernate.Hibernate;
+
+import java.util.Objects;
+
+@Getter
+@Setter
+@ToString
+@RequiredArgsConstructor
+@Builder(toBuilder = true)
+@AllArgsConstructor
+@NoArgsConstructor
+
+@Table
+@Entity
+public class Vote {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "vote_type")
+    @Enumerated(EnumType.STRING)
+    private VoteType voteType;
+
+    @NotNull
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", referencedColumnName = "id")
+    @ToString.Exclude
+    private Post post;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
+    @ToString.Exclude
+    private User user;
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
+        Vote vote = (Vote) o;
+        return id != null && Objects.equals(id, vote.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
+}
