+++ src/main/java/com/eacuamba/dev/reddit_clone_using_angular_spring/domain/model/vote/VoteType.java
@@ -1,0 +1,13 @@
+package com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.vote;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public enum VoteType {
+    UP(1),
+    DOWN(-1);
+
+    private final Integer direction;
+}
