+++ src/main/java/com/eacuamba/dev/reddit_clone_using_angular_spring/domain/repository/VoteRepository.java
@@ -1,7 +1,6 @@
 package com.eacuamba.dev.reddit_clone_using_angular_spring.domain.repository;
 
-import com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.User;
-import com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.Vote;
+import com.eacuamba.dev.reddit_clone_using_angular_spring.domain.model.vote.Vote;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
