+++ src/main/java/com/eacuamba/dev/reddit_clone_using_angular_spring/domain/service/VoteService.java
@@ -1,0 +1,13 @@
+package com.eacuamba.dev.reddit_clone_using_angular_spring.domain.service;
+
+import com.eacuamba.dev.reddit_clone_using_angular_spring.domain.repository.VoteRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@RequiredArgsConstructor
+@Slf4j
+@Component
+public class VoteService {
+    private final VoteRepository voteRepository;
+}
